Make addon.py slim by extracting logic to plugin.py
